requirements.txt
Pillow==3.2.0
requests==2.9.1
